set up a seperate libstdc++ test called libtbx.check_libcpp
