+++ libtbx/env_config.py
@@ -3200,78 +3200,10 @@ def get_boost_library_with_python_version(name, libpath):
         return name_version
   return name
 
-def check_libcpp():
-
-  #______________________________________________________________________________
-  # Skip if not Linux
-  # TODO: this might also work on macOS
-  #
-  if not sys.platform.startswith("linux"):
-    return True
-  #
-  #------------------------------------------------------------------------------
-
-  import libtbx.env_config
-  from   libtbx             import easy_run
-  from   os                 import listdir
-
-  #______________________________________________________________________________
-  # get the libstd version that the python executable links against
-  #
-  python_libstdcxx_so = None
-  ldd_exec = easy_run.fully_buffered(command="ldd " + str(sys.executable))
-  for line in ldd_exec.stdout_lines:
-    if line.strip().startswith("libstdc++.so"):
-      python_libstdcxx_so = line.split()[0]
-      break
-
-  if python_libstdcxx_so is None:
-    return False
-  #
-  #------------------------------------------------------------------------------
-
-  #______________________________________________________________________________
-  # Go through all installed *.so's and ensure that they are linking to the same
-  # libstdc++ as python executable
-  #
-  lib_dir = join(
-      libtbx.env.lib_path._anchor._path, libtbx.env.lib_path.relocatable
-  )
-  libs = [join(lib_dir, name) for name in listdir(lib_dir) if name.endswith(".so")]
-  for lib in libs:
-    ldd_lib = easy_run.fully_buffered(command="ldd " + str(lib))
-    for line in ldd_lib.stdout_lines:
-      if line.strip().startswith("libstdc++.so"):
-        lib_libstdcxx_so = line.split()[0]
-        if lib_libstdcxx_so != python_libstdcxx_so:
-          raise SystemError(
-            "FATAL: libstdc++.so mismatch\n"
-            "sys.executable: " + str(sys.executable) + "\n"
-            "lib_file:" + str(lib)
-          )
-        break
-  #
-  #------------------------------------------------------------------------------
-
-  return True
-
 
 if (__name__ == "__main__"):
   if (len(sys.argv) == 2 and sys.argv[1] == "__libtbx_refresh__"):
     unpickle().refresh()
-
-    def is_true(x):
-      yes = {"1", "on", "true"}
-      return (x.strip().lower() in yes)
-
-    if is_true(os.environ.get("CCTBX_CHECK_LDD", "True")):
-      print(
-          "Checking that extension modules and python executable link to same "
-          "version of libstdc++"
-      )
-      print("Set CCTBX_CHECK_LDD=false to disable this step")
-      check_libcpp()
-
   else:
     warm_start(sys.argv)
   print("Done.")
